Refactor: Add `lista_servicios` to `Ver_Reservation_Frame` and `init_ver_reservaciones`
This commit refactors the `ver_reservaciones.py` module by adding a `lista_servicios` parameter to both the `Ver_Reservation_Frame` class and the `init_ver_reservaciones` function. This allows the reservation view to display the list of available services.

Additionally, the `boton_reservaciones` in `home.py` is updated to pass the `lista_servicios` to `init_ver_reservaciones`.